updated project settings panel + other refactoring
