+++ Vortex/src/Vortex/Renderer/ParticleSystem/ParticleSystem.cpp
@@ -138,7 +138,6 @@ namespace Vortex {
 
 	void ParticleSystem::OnGuiRender()
 	{
-		
 	}
 
 }
